#20: Use graphql sdl in flutter
