+++ frontend/build.yaml
@@ -3,12 +3,12 @@ targets:
     sources:
       - lib/**
       - graphql_queries/**
-      - graphql.schema.json
+      - schema.graphql
     builders:
       artemis:
         options:
           schema_mapping:
-            - schema: graphql.schema.json
+            - schema: schema.graphql
               queries_glob: graphql_queries/*.graphql
               output: lib/graphql/graphql_api.dart
           scalar_mapping:
