+++ frontend/pubspec.yaml
@@ -29,10 +29,12 @@ dependencies:
   permission_handler: ^5.0.1+1
 
   # graphQL type generation
-  path_provider: ^1.5.1
-  equatable: ^1.0.2
-  json_serializable: ^3.2.3
-  gql: 0.12.0
+  artemis: '>=6.0.0 <7.0.0' # only if you're using ArtemisClient!
+  json_annotation: ^3.1.0
+  equatable: ^1.2.5
+  meta: '>=1.0.0 <2.0.0' # only if you have non nullable fields
+  gql: '>=0.12.3 <1.0.0'
+
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -41,8 +43,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.7.2
-  artemis: ^2.1.4
+  artemis: '>=6.0.0 <7.0.0'
+  build_runner: ^1.10.4
+  json_serializable: ^3.5.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
